initial commit 3

+++ src/main/java/org/example/camerarentweb/repositories/UserRepository.java
@@ -4,10 +4,12 @@ import org.example.camerarentweb.entities.User;
 import org.example.camerarentweb.entities.UserRole;
 
 import java.util.List;
+import java.util.Optional;
 
 public interface UserRepository extends BaseRepository<User> {
 
     User findByPhoneNumber(String phoneNumber);
+    //Optional<User> findByPhoneNumber(String phoneNumber);
 
     User findByEmail(String email);
 
@@ -23,7 +25,7 @@ public interface UserRepository extends BaseRepository<User> {
 
     User findByEmailAndPassword(String email, String password);
 
-    void activateUser(Long id);
+    void activateUser(int id);
 
-    void deactivateUser(Long id);
+    void deactivateUser(int id);
 }
